Splitting location logic
