+++ spiffworkflow-backend/src/spiffworkflow_backend/services/reference_cache_service.py
@@ -5,6 +5,7 @@ from sqlalchemy import insert
 from spiffworkflow_backend.models.cache_generation import CacheGenerationModel
 from spiffworkflow_backend.models.db import db
 from spiffworkflow_backend.models.reference_cache import ReferenceCacheModel
+from spiffworkflow_backend.services.upsearch_service import UpsearchService
 
 
 class ReferenceCacheService:
@@ -37,7 +38,7 @@ class ReferenceCacheService:
         cache_generation = CacheGenerationModel.newest_generation_for_table("reference_cache")
         if cache_generation is None:
             return None
-        locations = cls.upsearch_locations(location)
+        locations = UpsearchService.upsearch_locations(location)
         references = (
             ReferenceCacheModel.query.filter_by(
                 identifier=identifier,
@@ -54,13 +55,3 @@ class ReferenceCacheService:
             return reference.relative_location  # type: ignore
 
         return None
-
-    @classmethod
-    def upsearch_locations(cls, location: str) -> list[str]:
-        locations = []
-
-        while location != "":
-            locations.append(location)
-            location = os.path.dirname(location)
-
-        return locations
