+++ spiffworkflow-backend/src/spiffworkflow_backend/services/upsearch_service.py
@@ -1,0 +1,14 @@
+import os
+
+class UpsearchService:
+    @classmethod
+    def upsearch_locations(cls, location: str) -> list[str]:
+        locations = []
+
+        while location != "":
+            locations.append(location)
+            location = os.path.dirname(location)
+
+        return locations
+
+
