+++ spiffworkflow-backend/tests/spiffworkflow_backend/unit/test_reference_cache_service.py
@@ -41,17 +41,6 @@ def with_loaded_reference_cache(app: Flask, with_db_and_bpmn_file_cleanup: None)
 
 
 class TestReferenceCacheService(BaseTest):
-    def test_upsearch_locations(
-        self,
-    ) -> None:
-        locations = ReferenceCacheService.upsearch_locations("misc/jonjon/generic-data-store-area/test-level-2")
-        assert locations == [
-            "misc/jonjon/generic-data-store-area/test-level-2",
-            "misc/jonjon/generic-data-store-area",
-            "misc/jonjon",
-            "misc",
-        ]
-
     def test_can_find_data_store_in_current_location(self, with_loaded_reference_cache: None) -> None:
         location = ReferenceCacheService.upsearch(
             "misc/jonjon/generic-data-store-area/test-level-1", "contacts_datastore", "data_store"
